Nộp bài lab3

+++ security/list_users.php
@@ -1,7 +1,5 @@
 <?php
-// Start the session
 session_start();
-
 require_once 'models/UserModel.php';
 $userModel = new UserModel();
 
@@ -21,7 +19,7 @@ $users = $userModel->getUsers($params);
 <body>
     <?php include 'views/header.php'?>
     <div class="container">
-        <?php if (!empty($users)) {?>
+        <?php if (!empty($users)) { ?>
             <div class="alert alert-warning" role="alert">
                 List of users! <br>
                 Hacker: http://php.local/list_users.php?keyword=ASDF%25%22%3BTRUNCATE+banks%3B%23%23
@@ -37,18 +35,12 @@ $users = $userModel->getUsers($params);
                     </tr>
                 </thead>
                 <tbody>
-                    <?php foreach ($users as $user) {?>
+                    <?php foreach ($users as $user) { ?>
                         <tr>
                             <th scope="row"><?php echo $user['id']?></th>
-                            <td>
-                                <?php echo $user['name']?>
-                            </td>
-                            <td>
-                                <?php echo $user['fullname']?>
-                            </td>
-                            <td>
-                                <?php echo $user['type']?>
-                            </td>
+                            <td><?php echo $user['name']?></td>
+                            <td><?php echo $user['fullname']?></td>
+                            <td><?php echo $user['type']?></td>
                             <td>
                                 <a href="form_user.php?id=<?php echo $user['id'] ?>">
                                     <i class="fa fa-pencil-square-o" aria-hidden="true" title="Update"></i>
@@ -56,7 +48,7 @@ $users = $userModel->getUsers($params);
                                 <a href="view_user.php?id=<?php echo $user['id'] ?>">
                                     <i class="fa fa-eye" aria-hidden="true" title="View"></i>
                                 </a>
-                                <a href="delete_user.php?id=<?php echo $user['id'] ?>">
+                                <a href="delete_user.php?id=<?php echo $user['id'] ?>" onclick="return confirm('Bạn có chắc chắn muốn xóa người dùng này không?');">
                                     <i class="fa fa-eraser" aria-hidden="true" title="Delete"></i>
                                 </a>
                             </td>
@@ -64,7 +56,7 @@ $users = $userModel->getUsers($params);
                     <?php } ?>
                 </tbody>
             </table>
-        <?php }else { ?>
+        <?php } else { ?>
             <div class="alert alert-dark" role="alert">
                 This is a dark alert—check it out!
             </div>
